Refactor the folders Change AccountType.class to UserAccountType.class
