+++ src/main/java/com/school/project/model/entities/Lesson.java
@@ -1,8 +1,4 @@
-package com.school.project.model.lesson;
-
-import com.school.project.model.lesson.Group;
-import com.school.project.model.module.Subject;
-import com.school.project.model.user.User;
+package com.school.project.model.entities;
 
 import java.util.Date;
