+++ src/main/java/com/school/project/model/types/UserAccountType.java
@@ -1,0 +1,11 @@
+package com.school.project.model.types;
+
+public enum UserAccountType {
+    TEACHER(1, "Teacher"),
+    ADMIN(2, "Admin"),
+    STUDENT(3, "Student");
+
+    UserAccountType(Integer id, String name) {
+    }
+
+}
