--- src/main/java/com/school/project/model/user/AccountType.java
@@ -1,11 +1,0 @@
-package com.school.project.model.user;
-
-public enum AccountType {
-    TEACHER(1, "Teacher"),
-    ADMIN(2, "Admin"),
-    STUDENT(3, "Student");
-
-    AccountType(Integer id, String name) {
-    }
-
-}
