+++ src/main/java/com/school/project/repository/UserAccountRepository.java
@@ -1,6 +1,6 @@
 package com.school.project.repository;
 
-import com.school.project.model.user.UserAccount;
+import com.school.project.model.entities.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
