+++ src/main/java/com/school/project/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.school.project.repository;
 
-import com.school.project.model.user.User;
+import com.school.project.model.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
