+++ src/main/java/com/school/project/service/UserAccountService.java
@@ -1,9 +1,9 @@
 package com.school.project.service;
 
-import com.school.project.model.user.AccountType;
-import com.school.project.model.user.User;
-import com.school.project.model.user.UserAccount;
+import com.school.project.model.types.UserAccountType;
+import com.school.project.model.entities.User;
+import com.school.project.model.entities.UserAccount;
 
 public interface UserAccountService {
-    UserAccount createUserAccount(User user, AccountType accountType);
+    UserAccount createUserAccount(User user, UserAccountType userAccountType);
 }
