+++ src/main/java/com/school/project/service/impl/UserServiceImpl.java
@@ -2,8 +2,8 @@ package com.school.project.service.impl;
 
 import com.school.project.dto.UserDto;
 import com.school.project.exception.UserNotFoundException;
-import com.school.project.model.user.AccountType;
-import com.school.project.model.user.User;
+import com.school.project.model.types.UserAccountType;
+import com.school.project.model.entities.User;
 import com.school.project.repository.UserAccountRepository;
 import com.school.project.repository.UserRepository;
 import com.school.project.service.UserAccountService;
@@ -26,7 +26,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public User create(User user) {
         userRepository.save(user);
-        userAccountService.createUserAccount(user, AccountType.STUDENT);
+        userAccountService.createUserAccount(user, UserAccountType.STUDENT);
         return user;
     }
